Put header background image style inline
By placing the CSS property which defines the background image for the
header on the long-form inside of an inline style attribute, it is now
more clear as to how the header/banner/image may be dynamically changed
from within the view construct. This will make more sense for tailoring
the forms to be branded as though they were being offered directly by
some specific site which is part of the larger network of payday
websites.